requirements.txt
amqp==2.5.2
asgiref==3.2.5
billiard==3.6.3.0
celery==4.4.2
Django==3.0.4
django-crispy-forms==1.9.0
importlib-metadata==1.5.0
kombu==4.6.8
pkg-resources==0.0.0
psycopg2-binary==2.8.4
python-dotenv==0.12.0
pytz==2019.3
redis==3.4.1
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
